Issue #3403634 by SV: Refactor node access handler for event managers
